Update build_ngfd-plugin-droid-vibrator_package.sh
If local git repo exists do a pull instead of clone.
Added check if rpm packages were build successfully.

+++ build_ngfd-plugin-droid-vibrator_package.sh
@@ -9,9 +9,20 @@ PKG=ngfd-plugin-droid-vibrator
 SPEC=$PKG
 
 cd $MER_ROOT/devel/mer-hybris
-git clone https://github.com/mer-hybris/$PKG.git
+if [ ! -d $PKG ]; then
+	git clone https://github.com/mer-hybris/$PKG.git
+else
+	git pull
+fi
+
 cd $PKG
 mb2 -s rpm/$SPEC.spec -t $VENDOR-$DEVICE-armv7hl build
+
+if [ $(ls RPMS/*.rpm | wc -l) -eq 0 ]; then
+	echo "Error: No RPMs found for $PKG"
+	exit 1
+fi
+
 mkdir -p $ANDROID_ROOT/droid-local-repo/$DEVICE/$PKG/
 rm -f $ANDROID_ROOT/droid-local-repo/$DEVICE/$PKG/*.rpm
 mv RPMS/*.rpm $ANDROID_ROOT/droid-local-repo/$DEVICE/$PKG
